fix: check base permissions for ACO schemas

+++ packages/api-aco/src/folder/folder.gql.ts
@@ -1,6 +1,7 @@
 import { ErrorResponse, ListResponse } from "@webiny/handler-graphql/responses";
 import { GraphQLSchemaPlugin } from "@webiny/handler-graphql/plugins/GraphQLSchemaPlugin";
 
+import { checkPermissions } from "~/utils/checkPermissions";
 import { resolve } from "~/utils/resolve";
 
 import { AcoContext } from "~/types";
@@ -65,10 +66,12 @@ export const folderSchema = new GraphQLSchemaPlugin<AcoContext>({
     resolvers: {
         AcoQuery: {
             getFolder: async (_, { id }, context) => {
+                await checkPermissions(context);
                 return resolve(() => context.aco.folder.get(id));
             },
             listFolders: async (_, args: any, context) => {
                 try {
+                    await checkPermissions(context);
                     const [entries, meta] = await context.aco.folder.list(args);
                     return new ListResponse(entries, meta);
                 } catch (e) {
@@ -78,12 +81,15 @@ export const folderSchema = new GraphQLSchemaPlugin<AcoContext>({
         },
         AcoMutation: {
             createFolder: async (_, { data }, context) => {
+                await checkPermissions(context);
                 return resolve(() => context.aco.folder.create(data));
             },
             updateFolder: async (_, { id, data }, context) => {
+                await checkPermissions(context);
                 return resolve(() => context.aco.folder.update(id, data));
             },
             deleteFolder: async (_, { id }, context) => {
+                await checkPermissions(context);
                 return resolve(() => context.aco.folder.delete(id));
             }
         }
